+++ src/math.rs
@@ -1,20 +1,20 @@
-/// Safe arithmetic helpers for u128.
+/// Safe arithmetic helpers for u64.
 pub mod math {
-    /// Adds two u128 values, saturating on overflow.
+    /// Adds two u64 values, saturating on overflow.
     #[inline]
-    pub fn safe_add(a: u128, b: u128) -> u128 {
+    pub fn safe_add(a: u64, b: u64) -> u64 {
         a.saturating_add(b)
     }
 
-    /// Subtracts two u128 values, saturating on underflow.
+    /// Subtracts two u64 values, saturating on underflow.
     #[inline]
-    pub fn safe_sub(a: u128, b: u128) -> u128 {
+    pub fn safe_sub(a: u64, b: u64) -> u64 {
         a.saturating_sub(b)
     }
 
     /// Adds then subtracts values: base + plus - minus, safely.
     #[inline]
-    pub fn safe_add_sub(base: u128, plus: u128, minus: u128) -> u128 {
+    pub fn safe_add_sub(base: u64, plus: u64, minus: u64) -> u64 {
         base.saturating_add(plus).saturating_sub(minus)
     }
 }
@@ -26,7 +26,7 @@ mod tests_math {
     #[test]
     fn test_safe_add() {
         assert_eq!(safe_add(100, 50), 150);
-        assert_eq!(safe_add(u128::MAX, 1), u128::MAX);
+        assert_eq!(safe_add(u64::MAX, 1), u64::MAX);
     }
 
     #[test]
